+++ counsel-back/src/main/java/io/notfound/counsel_back/board/entity/PostComment.java
@@ -1,28 +1,26 @@
-package io.notfound.counsel_back.consultation.entity;
+package io.notfound.counsel_back.board.entity;
 
 import jakarta.persistence.*;
 import lombok.*;
+
+import java.time.LocalDateTime;
 
 @Entity
 @Getter
 @NoArgsConstructor
 @Builder
 @AllArgsConstructor
-public class Attachment {
+public class PostComment {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
-    private String fileName;
-    private String fileUrl;
+    private String comment;
+
+    private LocalDateTime createdAt = LocalDateTime.now();
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "post_id")
     private Post post;
-    
-    public void setPost(Post post) {
-        this.post = post;
-    }
-
 }
