+++ counsel-back/src/main/java/io/notfound/counsel_back/board/repository/AttachmentRepository.java
@@ -1,7 +1,7 @@
-package io.notfound.counsel_back.consultation.repository;
+package io.notfound.counsel_back.board.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import io.notfound.counsel_back.consultation.entity.Attachment;
+import io.notfound.counsel_back.board.entity.Attachment;
 
 public interface AttachmentRepository extends JpaRepository<Attachment, Long> { }
