+++ counsel-back/src/main/java/io/notfound/counsel_back/board/repository/PostCommentRepository.java
@@ -1,8 +1,8 @@
-package io.notfound.counsel_back.consultation.repository;
+package io.notfound.counsel_back.board.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import io.notfound.counsel_back.consultation.entity.PostComment;
+import io.notfound.counsel_back.board.entity.PostComment;
 
 public interface PostCommentRepository extends JpaRepository<PostComment, Long> { }
