+++ counsel-back/src/main/java/io/notfound/counsel_back/board/repository/PostRepository.java
@@ -1,6 +1,7 @@
-package io.notfound.counsel_back.consultation.repository;
+package io.notfound.counsel_back.board.repository;
 
-import io.notfound.counsel_back.consultation.entity.Post;
 import org.springframework.data.jpa.repository.JpaRepository;
+
+import io.notfound.counsel_back.board.entity.Post;
 
 public interface PostRepository extends JpaRepository<Post, Long> { }
